Fix:#55 - place getDaysInMonth to utils
